fix: planstate 분리
